\name{diabetes}
\docType{data}
\alias{diabetes}
\title{Diabetes Data}
\description{
Diabetic retinopathy: how strongly are times to blindness of a treated and an untreated eye correlated in patients suffering from diabetic retinopathy? The analysis is based on a sample of n=197 paired failure times (censoring 73\% and 49\% for the treated and untreated eyes, respectively) described by Huster, Brookmeyer, and Self (1989). Both eyes of an individual are observed for the same time, and therefore dots on the diagonal generally indicate pairs of censored times. }
\usage{data(diabetes)}
\format{A data.frame containing 197 rows.}
\source{\url{https://www.mayo.edu/research/documents/diabeteshtml/DOC-10027460/}}
\references{
Huster WJ, Brookmeyer R, Self SG. Modelling paired survival data with covariates. Biometrics 1989; 45:145-156.
}
\keyword{datasets}